requirements.txt
fastapi
uvicorn
python-multipart
Pillow
google-genai
mcp
aiohttp
